--- programs/anchor-simple-fundraiser/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "anchor-simple-fundraiser"
-version = "0.1.0"
-description = "Created with Anchor"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "lib"]
-name = "anchor_simple_fundraiser"
-
-[features]
-default = []
-cpi = ["no-entrypoint"]
-no-entrypoint = []
-no-idl = []
-no-log-ix-name = []
-idl-build = ["anchor-lang/idl-build"]
-
-[dependencies]
-anchor-lang = { version = "0.30.1", features = ["init-if-needed"] }
