--- programs/anchor-simple-fundraiser/src/constants.rs
@@ -1,1 +1,0 @@
-pub const ANCHOR_DISCRIMINATOR: usize = 8;
