--- programs/anchor-simple-fundraiser/src/instructions/mod.rs
@@ -1,9 +1,0 @@
-pub mod initialize;
-pub mod contribute;
-pub mod checker;
-pub mod refund;
-
-pub use initialize::*;
-pub use contribute::*;
-pub use checker::*;
-pub use refund::*;
