--- programs/anchor-simple-fundraiser/src/state/contributor.rs
@@ -1,7 +1,0 @@
-use anchor_lang::prelude::*;
-
-#[account]
-#[derive(InitSpace)]
-pub struct Contributor {
-    pub amount: u64,
-}
