--- programs/anchor-simple-fundraiser/src/state/mod.rs
@@ -1,7 +1,0 @@
-pub mod fundraiser;
-pub mod contributor;
-pub mod vault;
-
-pub use fundraiser::*;
-pub use contributor::*;
-pub use vault::*;
