+++ programs/capstone/src/constants.rs
@@ -1,0 +1,1 @@
+pub const ANCHOR_DISCRIMINATOR: usize = 8;
