+++ programs/capstone/src/error.rs
@@ -6,16 +6,8 @@ pub enum FundraiserError {
     TargetNotMet,
     #[msg("The amount to raise has been achieved")]
     TargetMet,
-    #[msg("The contribution is too big")]
-    ContributionTooBig,
-    #[msg("The contribution is too small")]
-    ContributionTooSmall,
-    #[msg("The maximum amount to contribute has been reached")]
-    MaximumContributionsReached,
     #[msg("The fundraiser has not ended yet")]
     FundraiserNotEnded,
     #[msg("The fundraiser has ended")]
     FundraiserEnded,
-    #[msg("Invalid total amount. i should be bigger than 3")]
-    InvalidAmount
 }
