+++ programs/capstone/src/instructions/checker.rs
