+++ programs/capstone/src/instructions/mod.rs
@@ -1,9 +1,9 @@
 pub mod initialize;
 pub mod contribute;
-pub mod deposit;
+pub mod checker;
 pub mod refund;
 
 pub use initialize::*;
 pub use contribute::*;
-pub use deposit::*;
+pub use checker::*;
 pub use refund::*;
