+++ programs/capstone/src/instructions/refund.rs
@@ -1,0 +1,82 @@
+use anchor_lang::{prelude::*, system_program::{transfer, Transfer}};
+
+use crate::{
+    state::{
+        Contributor,
+        Fundraiser,
+        Vault,
+    },
+    FundraiserError,
+};
+
+#[derive(Accounts)]
+pub struct Refund<'info> {
+    #[account(mut)]
+    pub contributor: Signer<'info>,
+    pub maker: SystemAccount<'info>,
+    #[account(
+        mut,
+        seeds = [b"fundraiser", maker.key().as_ref()],
+        bump = fundraiser.bump,
+    )]
+    pub fundraiser: Account<'info, Fundraiser>,
+    #[account(
+        mut,
+        seeds = [b"contributor", fundraiser.key().as_ref(), contributor.key().as_ref()],
+        bump,
+        close = contributor,
+    )]
+    pub contributor_account: Account<'info, Contributor>,
+    #[account(
+        mut,
+    )]
+    pub vault: Account<'info, Vault>,
+    pub system_program: Program<'info, System>,
+}
+
+impl<'info> Refund<'info> {
+    pub fn refund(&mut self) -> Result<()> {
+
+        // Check if the fundraising duration has been reached
+        let current_time = Clock::get()?.unix_timestamp;
+ 
+        require!(
+            self.fundraiser.deadline <= current_time,
+            FundraiserError::FundraiserNotEnded
+        );
+
+        require!(
+            self.fundraiser.current_amount < self.fundraiser.amount_to_raise,
+            FundraiserError::TargetMet
+        );
+
+        // Transfer the funds back to the contributor
+        // CPI to the token program to transfer the funds
+        let cpi_program = self.system_program.to_account_info();
+
+        // Transfer the funds from the vault to the contributor
+        let cpi_accounts = Transfer {
+            from: self.vault.to_account_info(),
+            to: self.contributor_account.to_account_info(),
+        };
+
+        // Signer seeds to sign the CPI on behalf of the fundraiser account
+        let signer_seeds: [&[&[u8]]; 1] = [&[
+            b"fundraiser".as_ref(),
+            self.maker.to_account_info().key.as_ref(),
+            &[self.fundraiser.bump],
+        ]];
+
+        // CPI context with signer since the fundraiser account is a PDA
+        let cpi_ctx = CpiContext::new_with_signer(cpi_program, cpi_accounts, &signer_seeds);
+
+        // Transfer the funds from the vault to the contributor
+        transfer(cpi_ctx, self.contributor_account.amount)?;
+
+        // Update the fundraiser state by reducing the amount contributed
+        self.fundraiser.current_amount -= self.contributor_account.amount;
+ 
+
+        Ok(())
+    }
+}
