+++ programs/capstone/src/state/mod.rs
@@ -1,5 +1,7 @@
 pub mod fundraiser;
 pub mod contributor;
+pub mod vault;
 
 pub use fundraiser::*;
 pub use contributor::*;
+pub use vault::*;
