+++ programs/capstone/src/state/vault.rs
